BUGFIX: Fixed some warnings, got rid of a buffer overflow when reading a line from the SerialPort
git-svn-id: file:///home/srv/www/butt.se/pnp/svn/trunk@64 f672c3ff-8b41-4f22-a4ab-2adcb4f732c7

+++ MachineAPI/MachineController.cpp
@@ -216,7 +216,7 @@ bool MachineController::ValidateCommand(MachineCommand &cmd, MachineEvent *&vali
 	int zMin = 0;
 	int zMax = 10000;
 	float rMin = 0;
-	float rMax = 2*M_PI;
+	float rMax = (float)(2 * M_PI);
 
 	MachineState state = cmd.GetAfterState(currentState);
 	cout << "AfterState: x:" << state.GetX() << " y:" << state.GetY() << " z:" << state.GetZ() << endl;
